+++ packages/sources/css/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.60.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.60.1...@vtmn/css@0.60.2) (2022-03-09)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
+## [0.60.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.60.0...@vtmn/css@0.60.1) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
 # [0.60.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.59.9...@vtmn/css@0.60.0) (2022-03-04)