+++ packages/sources/css/src/components/indicators/loader/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.22](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-loader@0.2.21...@vtmn/css-loader@0.2.22) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-loader
+
+
+
+
+
 ## [0.2.21](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-loader@0.2.20...@vtmn/css-loader@0.2.21) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-loader