+++ packages/sources/css/src/components/indicators/price/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-price@0.2.10...@vtmn/css-price@0.2.11) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-price
+
+
+
+
+
 ## [0.2.10](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-price@0.2.9...@vtmn/css-price@0.2.10) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-price