+++ packages/sources/css/src/components/indicators/tag/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tag@0.4.1...@vtmn/css-tag@0.4.2) (2022-03-09)
+
+**Note:** Version bump only for package @vtmn/css-tag
+
+
+
+
+
+## [0.4.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tag@0.4.0...@vtmn/css-tag@0.4.1) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-tag
+
+
+
+
+
 # [0.4.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tag@0.3.10...@vtmn/css-tag@0.4.0) (2022-03-04)