+++ packages/sources/css/src/components/navigation/search/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-search@0.4.11...@vtmn/css-search@0.4.12) (2022-03-09)
+
+**Note:** Version bump only for package @vtmn/css-search
+
+
+
+
+
+## [0.4.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-search@0.4.10...@vtmn/css-search@0.4.11) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-search
+
+
+
+
+
 ## [0.4.10](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-search@0.4.9...@vtmn/css-search@0.4.10) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-search