+++ packages/sources/css/src/utilities/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.5](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.3.4...@vtmn/css-utilities@0.3.5) (2022-03-07)
+
+**Note:** Version bump only for package @vtmn/css-utilities
+
+
+
+
+
 ## [0.3.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.3.3...@vtmn/css-utilities@0.3.4) (2022-03-03)
 
 **Note:** Version bump only for package @vtmn/css-utilities